refactor: share memory threat patterns
