+++ .github/workflows/ci.yml
@@ -63,7 +63,6 @@ jobs:
           install-video-tools: 'true'
       - run: pnpm --filter @amaster.ai/pi-storage prisma:generate
       - run: pnpm --filter @amaster.ai/pi-shared build
-      - run: pnpm --filter @amaster.ai/pi-memory build
       - run: pnpm test
 
   build:
