+++ packages/pi-memory-mem0/package.json
@@ -50,7 +50,6 @@
     "test": "vitest run src"
   },
   "dependencies": {
-    "@amaster.ai/pi-memory": "workspace:*",
     "@amaster.ai/pi-shared": "workspace:*",
     "mem0ai": "3.0.7"
   },
