+++ packages/pi-memory-mem0/tsconfig.json
@@ -5,7 +5,7 @@
     "rootDir": "src",
     "outDir": "dist"
   },
-  "references": [{ "path": "../shared" }, { "path": "../pi-memory" }],
+  "references": [{ "path": "../shared" }],
   "include": ["src/**/*.ts"],
   "exclude": ["src/**/*.test.ts"]
 }
