+++ packages/shared/package.json
@@ -15,6 +15,10 @@
       "types": "./dist/settings.d.ts",
       "default": "./dist/settings.js"
     },
+    "./threat-patterns": {
+      "types": "./dist/threat-patterns.d.ts",
+      "default": "./dist/threat-patterns.js"
+    },
     "./package.json": {
       "default": "./package.json"
     }
